refactor: Trains, TrainsManager moved to model package
